Update index.html and style.css I got the index.html part done
Reflection about css:
I was stuck on the css part for a while and I relied on the solution.css
Basically I wasn't able to control the divs at my liking
Just when I peeped at the solution.css, I realized that it was possible.

display: flex; automatically makes the divs in a row flow
this was my major problem since I was not able to align the left and right divs
I just needed to add display: flex; to the both which I don't know why I didn't think of it.

and I'm still having trouble familiarizing myself with the use of css rules because
the cascade rules of css are messing with my intuitive coding, with these I have to think instead of
just coding. And with lack of comprehension I just think wrongly.

But I'm glad I'm learning these use cases.

Still sticking to one of the quotes I read, that css only commands what you declared
and they're not coding against you.