Update index.html and style.css Second attempt and still committed a lot of mistakes
I was fixated on flex fixing everything but somehow the body type selector
fixes it with only text-align: center to the image as well.

max-width never occur to me.

I was wasting time declaring on many classes when I could just declare it with lesser classes

Still have a lot to go but I'm happy with the progress